EReaderHandleTransfer is u16

+++ include/ereader_helpers.h
@@ -24,7 +24,7 @@ struct EReaderTrainerHillSet
 bool8 EReader_IsReceivedDataValid(struct EReaderTrainerHillSet *buffer);
 bool32 TryWriteTrainerHill(struct EReaderTrainerHillSet *arg0);
 bool32 ReadTrainerHillAndValidate(void);
-int EReaderHandleTransfer(u8, u32, u32*, u32*);
+u16 EReaderHandleTransfer(u8, u32, u32*, u32*);
 void sub_81D3F9C(void);
 void sub_81D3FAC(void);
 void EReaderHelper_SaveRegsState(void);
